Move crossover implementations behind feature gate
